requirements.txt
ipykernel==7.2.0
pandas>=3.0.0
numpy>=2.4.0
Jinja2>=3.0.0
seaborn>=0.13.0
matplotlib>=3.10.0
scikit-learn>=1.6.0
tsfresh>=0.21.0
pyarrow>=23.0.0
sklearn-compat>=0.1.5
imbalanced-learn>=0.14.0
wandb>=0.16.0